removing bytes and fixed bytes
