+++ stylus-proc/src/types.rs
@@ -80,18 +80,22 @@ pub fn solidity_type_info(ty: &Type) -> (Cow<'static, str>, Cow<'static, str>) {
         Type::Address(_, _) => simple!(Address),
         Type::String(_) => simple!(String),
         Type::Bytes(_) => simple!(Bytes),
+        // TODO: Change this to simple type
         Type::FixedBytes(_, size) => (
             format!("stylus_sdk::abi::FixedBytesSolType<{size}>").into(),
             abi!("bytes[{size}]"),
         ),
+        // TODO: Review if simple type is possible now
         Type::Uint(_, size) => {
             let size = size.unwrap_or(NonZeroU16::new(256).unwrap());
             (path!("Uint<{size}>"), abi!("uint{size}"))
         }
+        // TODO: Review if simple type is possible now
         Type::Int(_, size) => {
             let size = size.unwrap_or(NonZeroU16::new(256).unwrap());
             (path!("Int<{size}>"), abi!("int{size}"))
         }
+        // TODO: Review if simple type is possible now
         Type::Array(ty) => {
             let (path, abi) = solidity_type_info(&ty.ty);
 
@@ -104,6 +108,7 @@ pub fn solidity_type_info(ty: &Type) -> (Cow<'static, str>, Cow<'static, str>) {
                 None => (path!("Array<{path}>"), abi!("{abi}[]")),
             }
         }
+        // TODO: Review if simple type is possible now
         Type::Tuple(tup) => {
             if tup.types.is_empty() {
                 ("()".into(), "()".into())
